비밀번호 변경 시 기존 비밀번호와 변경할 비밀번호가 같으면 예외를 발생

+++ bitgouel-api/src/main/kotlin/team/msg/domain/auth/service/AuthServiceImpl.kt
@@ -280,9 +280,12 @@ class AuthServiceImpl(
         if(!emailAuthentication.isAuthentication)
             throw UnAuthenticatedEmailException("아직 인증되지 않은 이메일입니다. info : [ email = ${request.email} ]")
 
-        emailAuthenticationRepository.delete(emailAuthentication)
-
         val encodedNewPassword = securityUtil.passwordEncode(request.newPassword)
+
+        if(securityUtil.isPasswordMatch(request.newPassword, user.password))
+            throw SameAsOldPasswordException("기존 비밀번호와 새 비밀번호가 일치합니다. info : [ password = ${request.newPassword} ]")
+
+        emailAuthenticationRepository.delete(emailAuthentication)
 
         val modifiedPasswordUser = User(
             id = user.id,
